Manage CategoryDropDown
feat: Manage CategoryDropDown by self or react router instead of redux state

The value change of CategoryDropDown is decided directly by self or url change. If there is a state for it, the state actually comes from those two sources. We cannot use the result to change its source. That means we can't use redux state to change the url since the url causes the state change.

- Remove the state currentCategory and all the related from redux state

- Make CategoryDropDown managed by self or the url change