PostPaper self-controlled component
feat: For editing post, use self-controlled component to be as a single unit for post and comment display

On the editing post page, it can appear many post paper, one post and multi comments. The total count of the paper is not fixed. In this case, it is hard to use redux state to control each paper's state. So use component itself to control its own state.

- Create PostPaper component as an unit

- Create PostCard as the child component of PostPaper and being controlled by it

- Post will include as many as PostPapers as it needs

- Remove useless code