Deleting post state
feat: Add a state in appStatus for deleting post action

When a post is deleting, it needs to consider the page after deleting, either redirecting to other pages or staying on the current page but showing some info for the deleting result. Here the app takes the latter. Under this circumstance, the page content should not disappear immediately when pressing the delete button. It should begin to respond after the asycn returns successfully. So here needs a state to coordinate with this process.

- Add the state and actions for monitoring the deleting process in appStatus

- Modify the previous deleting post action to make it change the redux state only when the async returns successfully

- Remove restoring action and the action monitor state from redux for no use anymore

- Add Backdrop component to display the view when deleting is processing

- Implement deleting on the post editing page

- Rename the loading bar monitoring state to avoid the conflict with its related actions